requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
pydantic
resend
